finished permission modal added group list
